+++ opa-evaluator/src/main/java/io/github/open_policy_agent/opa/bundle/Bundle.java
@@ -6,7 +6,7 @@ import io.github.open_policy_agent.opa.ir.policy.Policy;
 
 public class Bundle {
   public final Policy irPolicy;
-  public final Map<String, Object> manifest;
+  public final Manifest manifest;
   public final Map<String, String> rego;
 
   private Bundle(Builder builder) {
@@ -17,7 +17,7 @@ public class Bundle {
 
   public static class Builder {
     private final Map<String, String> rego = new HashMap<>();
-    private Map<String, Object> manifest;
+    private Manifest manifest;
     private Policy irPolicy;
 
     public Builder withRego(String path, String rego) {
@@ -25,7 +25,7 @@ public class Bundle {
       return this;
     }
 
-    public Builder withManifest(Map<String, Object> manifest) {
+    public Builder withManifest(Manifest manifest) {
       this.manifest = manifest;
       return this;
     }